include location's country and local date/time in message

+++ team-00/webflux/src/main/java/com/example/demo/entities/openweathermap/OpenWeatherMapResponse.java
@@ -18,7 +18,7 @@ public class OpenWeatherMapResponse {
     private Rain rain;
     private Snow snow;
     private Sys sys;
-    private Long timezone;
+    private Integer timezone;
     private Long visibility;
     private List<Weather> weather;
     private Wind wind;
